+++ ricaun.RevitTest.Command/Process/RevitTestProcessStart.cs
@@ -29,7 +29,13 @@ namespace ricaun.RevitTest.Command.Process
             if (!log) return this;
             return SetRevitArgument("log");
         }
-        public RevitTestProcessStart SetRevitVersion(int revitVersion) => SetRevitArgument("version", revitVersion);
+        public RevitTestProcessStart SetRevitVersion(string revitVersion)
+        {
+            if (string.IsNullOrWhiteSpace(revitVersion) == false)
+                SetRevitArgument("version", revitVersion);
+
+            return this;
+        }
         public RevitTestProcessStart SetRevitLanguage(string revitLanguage)
         {
             if (string.IsNullOrWhiteSpace(revitLanguage) == false)
@@ -128,7 +134,7 @@ namespace ricaun.RevitTest.Command.Process
             return Run(outputConsole, outputError);
         }
 
-        public Task RunReadTests(Action<string[]> actionTests, 
+        public Task RunReadTests(Action<string[]> actionTests,
             Action<string> consoleAction = null,
             Action<string> debugAction = null,
             Action<string> errorAction = null)